recommender-ai-service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
requests==2.31.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.9
pika==1.3.2
PyJWT==2.8.0
torch==2.0.1
sentence-transformers==2.2.2
chromadb==0.4.22
numpy==1.24.3
scikit-learn==1.3.0
